+++ packages/react/src/hooks/datasource/index.ts
@@ -1,4 +1,5 @@
 export * from "@/patterns/OneFilterPicker/types"
+export * from "./adaptDataAdapterToInfiniteScroll"
 export * from "./itemNeighbors"
 export * from "./types"
 export * from "./useData"